requirements.txt
aiohttp==3.6.3
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
discord.py==1.5.1
dnspython==1.15.0
dnspython3==1.15.0
idna==2.10
mcstatus==4.0.0
multidict==4.7.6
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
systemd==0.16.1
tzlocal==2.1
urllib3==1.25.11
yarl==1.5.1
